+++ packages/subgraph/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- update to use new plural `InsertNotes` event
 - fix typo in `handleJoinSplit`
 - rename `subtreeIndex` in tree insertion entities to `subtreeBatchOffset`
 - add tree insertion events